Reworked the Mixin plugin for a more dynamic compatibility system
